Revert "feat(bubu-log): move Siri shortcut setup to dedicated page"
This reverts commit 0938f2a32bcd4fab538680f18230b54389259bf1.
